Printing Compression ratio Final Commit

+++ src/main/Main.java
@@ -21,8 +21,8 @@ public class Main {
         StringBuilder binaryString = binaryGenerator.convertNormalStringToBinaryString();
         HuffmanGenerator huffmanGenerator = new HuffmanGenerator(userInput);
         StringBuilder huffmanString = huffmanGenerator.convertNormalStringToHuffmanString();
-        int lengthOfBinaryString = binaryString.length();
-        int lengthOfHuffmanString = huffmanString.length();
+        double lengthOfBinaryString = binaryString.length();
+        double lengthOfHuffmanString = huffmanString.length();
         System.out.println("The binary string has " + lengthOfBinaryString + " characters.");
         System.out.println("The huffman string has " + lengthOfHuffmanString + " characters.");
         System.out.println("Compression Ratio (Huffman/Binary): " + (lengthOfHuffmanString / lengthOfBinaryString) * 100 + "%");
